+++ src/analysis/Analyzer.jl
@@ -1,12 +1,13 @@
 module Analyzer
 
-export LSAnalyzer, inference_error_report_stack, inference_error_report_severity, initialize_target_modules!
+export LSAnalyzer, inference_error_report_stack, inference_error_report_severity, reset_report_target_modules!
+export UndefVarErrorReport, FieldErrorReport, BoundsErrorReport
 
 using Core.IR
 using JET.JETInterface
 using JET: JET, CC
 
-using ..JETLS: AnalysisEntry
+using ..JETLS: JETLS, AnalysisEntry
 using ..LSP
 
 # JETLS internal interface
@@ -29,57 +30,75 @@ inference_error_report_severity(@nospecialize report::JET.InferenceErrorReport) 
     inference_error_report_severity_impl(report)::DiagnosticSeverity.Ty
 
 """
-    InterpretationStateCache
-
-Internal state of `LSAnalyzer` that allows it to access the state of `LSInterpreter`.
-"""
-struct InterpretationStateCache
-    target_modules::Set{Module}
-    function InterpretationStateCache(;
-            target_modules = nothing
-        )
-        if target_modules === nothing
-            new(copy(empty_target_modules))
-        else
-            new(Set{Module}(target_modules))
-        end
-    end
-end
-
-const empty_target_modules = Set{Module}()
-
-"""
     LSAnalyzer <: AbstractAnalyzer
 
-This is a code analyzer specially designed for the language server.
+A code analyzer specially designed for the language server.
 It is implemented using the `JET.AbstractAnalyzer` framework,
 extending the base abstract interpretation performed by the Julia compiler
-to detect errors during analysis, along with analyzing types and effects.
-
-Currently, it analyzes the following errors:
-- `UndefVarErrorReport`: Reports undefined variables:
-  * [x] Reports undefined global variables
-  * [x] Reports undefined local variables
-  * [ ] Reports undefined static parameters
+to detect [`LSErrorReport`](@ref)s, along with analyzing types and effects.
 """
 struct LSAnalyzer <: ToplevelAbstractAnalyzer
     state::AnalyzerState
     analysis_token::AnalysisToken
-    cache::InterpretationStateCache
     method_table::CC.CachedMethodTable{CC.InternalMethodTable}
-    function LSAnalyzer(state::AnalyzerState, analysis_token::AnalysisToken, cache::InterpretationStateCache)
+
+    """
+        `LSAnalyzer.report_target_modules::::Union{Nothing,Set{Module}}`
+
+    Configures from which modules reports should be analyzed
+    - `report_target_modules === nothing`: Do not filter by module (used by tests)
+    - `report_target_modules::Set{Module}`: Only modules included in `report_target_modules`
+      will be subject to report analysis
+    """
+    report_target_modules::Union{Nothing,Set{Module}}
+
+    """
+        LSAnalyzer(state::AnalyzerState, analysis_token::AnalysisToken, report_target_modules::Set{Module})
+
+    Internal constructor of `LSAnalyzer`.
+    Used for both initial construction and creating a new [`LSAnalyzer`](@ref) from an existing one.
+    """
+    function LSAnalyzer(
+            state::AnalyzerState, analysis_token::AnalysisToken, report_target_modules::Union{Nothing,Set{Module}}
+        )
         method_table = CC.CachedMethodTable(CC.InternalMethodTable(state.world))
-        return new(state, analysis_token, cache, method_table)
+        return new(state, analysis_token, method_table, report_target_modules)
     end
 end
+
+"""
+    LSAnalyzer(entry::AnalysisEntry, state::AnalyzerState; report_target_modules=missing)
+        -> analyzer::LSAnalyzer
+
+Internal utility constructor for [`analyzer::LSAnalyzer`](@ref), which initializes
+`analyzer.report_target_modules` and `analyzer.analysis_token`.
+All new analysis entries should construct `LSAnalyzer` through this method.
+
+`report_target_modules` controls which modules are analyzed:
+- `missing`: Use the module list incrementally updated by [`reset_report_target_modules!`](@ref)
+- `nothing`: Do not filter by module (used by tests)
+- Otherwise: Create `Set{Module}` from an iterator of modules (used by tests).
+  Note that test code may also be updated by `reset_report_target_modules!`.
+"""
 function LSAnalyzer(
         @nospecialize(entry::AnalysisEntry), state::AnalyzerState;
-        target_modules = nothing
+        report_target_modules = missing
     )
-    analysis_cache_key = JET.compute_hash(entry, state.inf_params)
+    # N.B. Separate the cache by the identity of `report_target_modules`.
+    # The case `report_target_modules === missing` is a special case.
+    # In this case, `report_target_modules` is tracked incrementally using `reset_report_target_modules!`,
+    # but this is only used by the legacy analysis mode, and in that mode,
+    # analysis is performed by creating anonymous modules that essentially represent the same module,
+    # so there is no need to separate the cache by the identity of those anonymous modules
+    report_target_modules_hash = objectid(report_target_modules)
+    analysis_cache_key = JET.compute_hash(entry, state.inf_params, report_target_modules_hash)
     analysis_token = @lock LS_ANALYZER_CACHE_LOCK get!(AnalysisToken, LS_ANALYZER_CACHE, analysis_cache_key)
-    cache = InterpretationStateCache(; target_modules)
-    return LSAnalyzer(state, analysis_token, cache)
+    if report_target_modules === missing
+        report_target_modules = Set{Module}()
+    elseif report_target_modules !== nothing
+        report_target_modules = Set{Module}(report_target_modules)
+    end
+    return LSAnalyzer(state, analysis_token, report_target_modules)
 end
 
 # AbstractInterpreter API
@@ -100,7 +119,7 @@ CC.ipo_lattice(::LSAnalyzer) =
 JETInterface.AnalyzerState(analyzer::LSAnalyzer) = analyzer.state
 function JETInterface.AbstractAnalyzer(analyzer::LSAnalyzer, state::AnalyzerState)
     # XXX `analyzer.analysis_token` doesn't respect changes in `state.inf_params`
-    return LSAnalyzer(state, analyzer.analysis_token, analyzer.cache)
+    return LSAnalyzer(state, analyzer.analysis_token, analyzer.report_target_modules)
 end
 JETInterface.AnalysisToken(analyzer::LSAnalyzer) = analyzer.analysis_token
 
@@ -110,12 +129,13 @@ const LS_ANALYZER_CACHE_LOCK = ReentrantLock()
 # internal API
 # ============
 
-function initialize_target_modules!(analyzer::LSAnalyzer, analyzed_files::Dict{String,JET.AnalyzedFileInfo})
-    target_modules = analyzer.cache.target_modules
-    empty!(target_modules)
+function reset_report_target_modules!(analyzer::LSAnalyzer, analyzed_files::Dict{String,JET.AnalyzedFileInfo})
+    report_target_modules = analyzer.report_target_modules
+    isnothing(report_target_modules) && return nothing
+    empty!(report_target_modules)
     for (_, analyzed_file_info) in analyzed_files
         for module_range_info in analyzed_file_info.module_range_infos
-            push!(target_modules, last(module_range_info))
+            push!(report_target_modules, last(module_range_info))
         end
     end
     nothing
@@ -125,8 +145,23 @@ end
 # =========
 
 function should_analyze(analyzer::LSAnalyzer, sv::CC.InferenceState)
-    target_modules = analyzer.cache.target_modules
-    return isempty(target_modules) || CC.frame_module(sv) ∈ target_modules
+    report_target_modules = analyzer.report_target_modules
+    return isnothing(report_target_modules) || CC.frame_module(sv) ∈ report_target_modules
+end
+
+# Many builtin functions are used in `getproperty(::Any,::Symbol)` or `getindex(::Tuple,::Int)`, etc.,
+# so analysis injection needs to be performed in the context of those methods (i.e. `Base`),
+# but `report_target_modules` does not necessarily include `Base`.
+# Therefore, we need to check the inference frame one level up, and if it is in `report_target_modules`,
+# we also need to analyze it.
+function should_analyze_for_builtins(analyzer::LSAnalyzer, sv::CC.InferenceState)
+    report_target_modules = analyzer.report_target_modules
+    isnothing(report_target_modules) && return 0
+    CC.frame_module(sv) ∈ report_target_modules && return 0
+    checkbounds(Bool, sv.callstack, sv.parentid) || return nothing
+    parent = sv.callstack[sv.parentid]
+    CC.frame_module(parent) ∈ report_target_modules && return 1
+    return nothing
 end
 
 # analysis injections
@@ -150,6 +185,31 @@ non-concrete call sites in a toplevel frame created by `JET.virtual_process`.
 """
 CC.bail_out_toplevel_call(::LSAnalyzer, ::CC.InferenceState) = false
 
+@static if VERSION ≥ v"1.12.2"
+function CC.concrete_eval_eligible(
+        analyzer::LSAnalyzer, @nospecialize(f), result::CC.MethodCallResult,
+        arginfo::CC.ArgInfo, sv::CC.InferenceState
+    )
+    res = @invoke CC.concrete_eval_eligible(
+        analyzer::ToplevelAbstractAnalyzer, f::Any, result::CC.MethodCallResult,
+        arginfo::CC.ArgInfo, sv::CC.InferenceState)
+    # Ensure that semi-concrete interpretation is definitely disabled to prevent it from occurring
+    return res === :concrete_eval ? res : :none
+end
+# This overload disables concrete evaluation ad-hoc when concrete evaluation returns `Bottom`
+# (i.e., when an error occurs during concrete evaluation) and falls back to constant propagation
+# to enable error reporting
+function CC.concrete_eval_call(
+        analyzer::LSAnalyzer, @nospecialize(f), result::CC.MethodCallResult, arginfo::CC.ArgInfo,
+        sv::CC.InferenceState, invokecall::Union{CC.InvokeCall,Nothing}
+    )
+    res = @invoke CC.concrete_eval_call(
+        analyzer::ToplevelAbstractAnalyzer, f::Any, result::CC.MethodCallResult, arginfo::CC.ArgInfo,
+        sv::CC.InferenceState, invokecall::Union{CC.InvokeCall,Nothing})
+    return res.rt === Union{} ? nothing : res
+end
+end # @static if VERSION ≥ v"1.12.2"
+
 # TODO Better to factor out and share it with `JET.JETAnalyzer`
 function CC.abstract_eval_globalref(analyzer::LSAnalyzer,
     g::GlobalRef, saw_latestworld::Bool, sv::CC.InferenceState)
@@ -157,14 +217,39 @@ function CC.abstract_eval_globalref(analyzer::LSAnalyzer,
         return CC.RTEffects(Any, Any, CC.generic_getglobal_effects)
     end
     (valid_worlds, ret) = CC.scan_leaf_partitions(analyzer, g, sv.world) do analyzer::LSAnalyzer, binding::Core.Binding, partition::Core.BindingPartition
-        if should_analyze(analyzer, sv)
+        offset = should_analyze_for_builtins(analyzer, sv)
+        if offset !== nothing
             if partition.min_world ≤ sv.world.this ≤ partition.max_world # XXX This should probably be fixed on the Julia side
-                report_undef_global_var!(analyzer, sv, binding, partition)
+                report_undef_global_var!(analyzer, sv, binding, partition, offset)
             end
         end
         CC.abstract_eval_partition_load(analyzer, binding, partition)
     end
     CC.update_valid_age!(sv, valid_worlds)
+    return ret
+end
+
+function CC.builtin_tfunction(analyzer::LSAnalyzer,
+    @nospecialize(f), argtypes::Vector{Any}, sv::CC.InferenceState) # `AbstractAnalyzer` isn't overloaded on `return_type`
+    ret = @invoke CC.builtin_tfunction(analyzer::ToplevelAbstractAnalyzer,
+        f::Any, argtypes::Vector{Any}, sv::CC.InferenceState)
+    if f === fieldtype
+        # the valid widest possible return type of `fieldtype_tfunc` is `Union{Type,TypeVar}`
+        # because fields of unwrapped `DataType`s can legally be `TypeVar`s,
+        # but this will lead to lots of false positive `MethodErrorReport`s for inference
+        # with accessing to abstract fields since most methods don't expect `TypeVar`
+        # (e.g. `@report_call readuntil(stdin, 'c')`)
+        # JET.jl further widens this case to `Any` and give up further analysis rather than
+        # trying hard to do sound and noisy analysis
+        # xref: https://github.com/JuliaLang/julia/pull/38148
+        if ret === Union{Type, TypeVar}
+            ret = Any
+        end
+    end
+    offset = should_analyze_for_builtins(analyzer, sv)
+    if offset !== nothing
+        report_builtin_error!(analyzer, sv, f, argtypes, ret, offset)
+    end
     return ret
 end
 
@@ -184,12 +269,28 @@ end
 # analysis
 # ========
 
+"""
+    LSErrorReport <: InferenceErrorReport
+
+Abstract type for error reports analyzed by [`LSAnalyzer`](@ref).
+
+Subtypes:
+- `UndefVarErrorReport`: Undefined variables (global, local, static parameters[^unimplemented])
+- `FieldErrorReport`: Access to non-existent struct fields
+- `BoundsErrorReport`: Out-of-bounds field access by index
+- `MethodErrorReport`: Method dispatch errors[^unimplemented]
+
+[^unimplemented]: Currently unimplemented.
+"""
+abstract type LSErrorReport <: InferenceErrorReport end
+
 # UndefVarErrorReport
 # -------------------
 
-@jetreport struct UndefVarErrorReport <: InferenceErrorReport
+@jetreport struct UndefVarErrorReport <: LSErrorReport
     var::Union{GlobalRef,TypeVar,Symbol}
     maybeundef::Bool
+    vst_offset::Int
 end
 function JETInterface.print_report_message(io::IO, r::UndefVarErrorReport)
     var = r.var
@@ -208,12 +309,14 @@ function JETInterface.print_report_message(io::IO, r::UndefVarErrorReport)
         end
     end
 end
-inference_error_report_stack_impl(r::UndefVarErrorReport) = length(r.vst):-1:1
+inference_error_report_stack_impl(r::UndefVarErrorReport) = (length(r.vst)-r.vst_offset):-1:1
 inference_error_report_severity_impl(r::UndefVarErrorReport) =
     r.maybeundef ? DiagnosticSeverity.Information : DiagnosticSeverity.Warning
 
-function report_undef_global_var!(analyzer::LSAnalyzer,
-    sv::CC.InferenceState, binding::Core.Binding, partition::Core.BindingPartition)
+function report_undef_global_var!(
+        analyzer::LSAnalyzer, sv::CC.InferenceState, binding::Core.Binding, partition::Core.BindingPartition,
+        offset::Int
+    )
     gr = binding.globalref
     # TODO use `abstract_eval_isdefinedglobal` for respecting world age
     if @invokelatest isdefinedglobal(gr.mod, gr.name)
@@ -230,7 +333,122 @@ function report_undef_global_var!(analyzer::LSAnalyzer,
         binding_state.maybeundef || return false
         maybeundef = true
     end
-    add_new_report!(analyzer, sv.result, UndefVarErrorReport(sv, gr, maybeundef))
+    add_new_report!(analyzer, sv.result, UndefVarErrorReport(sv, gr, maybeundef, offset))
+    return true
+end
+
+@jetreport struct FieldErrorReport <: LSErrorReport
+    @nospecialize type
+    field::Symbol
+    vst_offset::Int
+end
+function JETInterface.print_report_message(io::IO, r::FieldErrorReport)
+    typ = r.type::Union{UnionAll,DataType}
+    flds = join(map(n->"`$n`", fieldnames(typ)), ", ")
+    if typ <: Tuple
+        typ = Tuple # reproduce base error message
+    end
+    @static if VERSION ≥ v"1.12.0-beta4.14"
+        # JuliaLang/julia#58507
+        typ = Base.unwrap_unionall(typ)::DataType
+        tname = string(typ.name.wrapper)
+    else
+        tname = nameof(typ)
+    end
+    return print(io, lazy"FieldError: type $tname has no field `$(r.field)`, available fields: $flds")
+end
+inference_error_report_stack_impl(r::FieldErrorReport) = (length(r.vst)-r.vst_offset):-1:1
+inference_error_report_severity_impl(::FieldErrorReport) = DiagnosticSeverity.Warning
+
+@jetreport struct BoundsErrorReport <: LSErrorReport
+    @nospecialize a
+    i::Int
+    vst_offset::Int
+end
+JETInterface.print_report_message(io::IO, r::BoundsErrorReport) =
+    print(io, lazy"BoundsError: attempt to access $(r.a) at index [$(r.i)]")
+inference_error_report_stack_impl(r::BoundsErrorReport) = (length(r.vst)-r.vst_offset):-1:1
+inference_error_report_severity_impl(::BoundsErrorReport) = DiagnosticSeverity.Warning
+
+function report_builtin_error!(
+        analyzer::LSAnalyzer, sv::CC.InferenceState, @nospecialize(f), argtypes::Vector{Any},
+        @nospecialize(ret), offset::Int
+    )
+    if ret === Union{}
+        if f === getfield
+            report_fieldaccess!(analyzer, sv, getfield, argtypes, offset)
+        elseif f === setfield!
+            report_fieldaccess!(analyzer, sv, setfield!, argtypes, offset)
+        elseif f === fieldtype
+            report_fieldaccess!(analyzer, sv, fieldtype, argtypes, offset)
+        end
+    end
+end
+
+const MODULE_SETFIELD_MSG = "cannot assign variables in other modules"
+type_error_msg(f, expected, actual) = (@nospecialize;
+    lazy"TypeError: in $f, expected $expected, got a value of type $actual")
+
+function report_fieldaccess!(
+        analyzer::LSAnalyzer, sv::CC.InferenceState, @nospecialize(f), argtypes::Vector{Any},
+        offset::Int
+    )
+    2 ≤ length(argtypes) ≤ 3 || return false
+
+    issetfield! = f === setfield!
+    obj, name = argtypes[1], argtypes[2]
+    s00 = CC.widenconst(obj)
+
+    if issetfield!
+        if !CC._mutability_errorcheck(s00)
+            # msg = lazy"setfield!: immutable struct of type $s00 cannot be changed"
+            # report = BuiltinErrorReport(sv, setfield!, msg, offset)
+            # add_new_report!(analyzer, sv.result, report)
+            return true
+        end
+    end
+
+    isa(name, Const) || return false
+    s = Base.unwrap_unionall(s00)
+    if CC.isType(s)
+        if f === fieldtype
+            # XXX this is a hack to share more code between `getfield`/`setfield!`/`fieldtype`
+            s00 = s = s.parameters[1]
+        elseif CC.isconstType(s)
+            s = (s00::DataType).parameters[1]
+        else
+            return false
+        end
+    end
+    isa(s, DataType) || return false
+    isabstracttype(s) && return false
+    if s <: Module
+        if issetfield!
+            # report = BuiltinErrorReport(sv, setfield!, MODULE_SETFIELD_MSG)
+            # add_new_report!(analyzer, sv.result, report, offset)
+            return true
+        end
+        nametyp = CC.widenconst(name)
+        if !CC.hasintersect(nametyp, Symbol)
+            # msg = type_error_msg(getglobal, Symbol, nametyp)
+            # report = BuiltinErrorReport(sv, getglobal, msg)
+            # add_new_report!(analyzer, sv.result, report, offset)
+            return true
+        end
+    end
+    fidx = CC._getfield_fieldindex(s, name)
+    if fidx !== nothing
+        nf = length(Base.datatype_fieldtypes(s))
+        1 ≤ fidx ≤ nf && return false
+    end
+
+    namev = (name::Const).val
+    objtyp = s
+    if namev isa Symbol
+        add_new_report!(analyzer, sv.result, FieldErrorReport(sv, objtyp, namev, offset))
+    elseif namev isa Int
+        add_new_report!(analyzer, sv.result, BoundsErrorReport(sv, objtyp, namev, offset))
+    else error("invalid field analysis") end
     return true
 end
 
@@ -261,7 +479,7 @@ function report_undefined_local_vars!(analyzer::LSAnalyzer, sv::CC.InferenceStat
         return false
     end
     maybeundef = vtype.typ !== Union{}
-    add_new_report!(analyzer, sv.result, UndefVarErrorReport(sv, name, maybeundef))
+    add_new_report!(analyzer, sv.result, UndefVarErrorReport(sv, name, maybeundef, #=vst_offset=#0))
     return true
 end
 
@@ -269,23 +487,24 @@ end
 # ===========
 
 # the entry constructor
-function LSAnalyzer(@nospecialize(entry::AnalysisEntry),
-                    world::UInt = Base.get_world_counter();
-                    target_modules = nothing,
-                    jetconfigs...)
+function LSAnalyzer(
+        @nospecialize(entry::AnalysisEntry), world::UInt = Base.get_world_counter();
+        report_target_modules = missing,
+        jetconfigs...
+    )
     jetconfigs = JET.kwargs_dict(jetconfigs)
     jetconfigs[:aggressive_constant_propagation] = true
     # Enable the `assume_bindings_static` option to terminate analysis a bit earlier when
     # there are undefined bindings detected. Note that this option will cause inference
     # cache inconsistency until JuliaLang/julia#40399 is merged. But the analysis cache of
-    # JETAnalyzer has the same problem already anyway, so enabling this option does not
+    # LSAnalyzer has the same problem already anyway, so enabling this option does not
     # make the situation worse.
     jetconfigs[:assume_bindings_static] = true
     state = AnalyzerState(world; jetconfigs...)
-    return LSAnalyzer(entry, state; target_modules)
+    return LSAnalyzer(entry, state; report_target_modules)
 end
 
-const LS_ANALYZER_CONFIGURATIONS = Set{Symbol}((:target_modules,))
+const LS_ANALYZER_CONFIGURATIONS = Set{Symbol}((:report_target_modules,))
 
 let valid_keys = JET.GENERAL_CONFIGURATIONS ∪ LS_ANALYZER_CONFIGURATIONS
     @eval JETInterface.valid_configurations(::LSAnalyzer) = $valid_keys