+++ src/types.jl
@@ -337,6 +337,8 @@ const TOPLEVEL_ERROR_CODE = "toplevel/error"
 const INFERENCE_UNDEF_GLOBAL_VAR_CODE = "inference/undef-global-var"
 const INFERENCE_UNDEF_LOCAL_VAR_CODE = "inference/undef-local-var"
 const INFERENCE_UNDEF_STATIC_PARAM_CODE = "inference/undef-static-param" # currently not reported
+const INFERENCE_FIELD_ERROR_CODE = "inference/field-error"
+const INFERENCE_BOUNDS_ERROR_CODE = "inference/bounds-error"
 const TESTRUNNER_TEST_FAILURE_CODE = "testrunner/test-failure"
 
 const ALL_DIAGNOSTIC_CODES = Set{String}(String[
@@ -349,6 +351,8 @@ const ALL_DIAGNOSTIC_CODES = Set{String}(String[
     INFERENCE_UNDEF_GLOBAL_VAR_CODE,
     INFERENCE_UNDEF_LOCAL_VAR_CODE,
     INFERENCE_UNDEF_STATIC_PARAM_CODE,
+    INFERENCE_FIELD_ERROR_CODE,
+    INFERENCE_BOUNDS_ERROR_CODE,
     TESTRUNNER_TEST_FAILURE_CODE,
 ])